Moving ErrorMessage and ErrorType to the exception package
