+++ aftership-sdk/src/main/java/com/aftership/sdk/endpoint/AfterShipEndpoint.java
@@ -3,8 +3,8 @@ package com.aftership.sdk.endpoint;
 import java.text.MessageFormat;
 import java.util.HashMap;
 import java.util.Map;
-import com.aftership.sdk.error.ErrorMessage;
-import com.aftership.sdk.error.ErrorType;
+import com.aftership.sdk.exception.ErrorMessage;
+import com.aftership.sdk.exception.ErrorType;
 import com.aftership.sdk.exception.SdkException;
 import com.aftership.sdk.model.AftershipResponse;
 import com.aftership.sdk.model.tracking.SlugTrackingNumber;
