+++ aftership-sdk/src/main/java/com/aftership/sdk/exception/ErrorType.java
@@ -1,4 +1,4 @@
-package com.aftership.sdk.error;
+package com.aftership.sdk.exception;
 
 import com.aftership.sdk.utils.StrUtils;
