+++ aftership-sdk/src/main/java/com/aftership/sdk/request/ApiRequestImpl.java
@@ -8,10 +8,10 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 import com.aftership.sdk.AfterShip;
-import com.aftership.sdk.error.ErrorMessage;
-import com.aftership.sdk.error.ErrorType;
 import com.aftership.sdk.exception.AftershipException;
 import com.aftership.sdk.exception.ApiException;
+import com.aftership.sdk.exception.ErrorMessage;
+import com.aftership.sdk.exception.ErrorType;
 import com.aftership.sdk.exception.RequestException;
 import com.aftership.sdk.model.AftershipResponse;
 import com.aftership.sdk.model.Meta;
