+++ aftership-sdk/src/test/java/com/aftership/sdk/error/ErrorMessageTest.java
@@ -2,6 +2,7 @@ package com.aftership.sdk.error;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import com.aftership.sdk.exception.ErrorMessage;
 import lombok.SneakyThrows;
 
 class ErrorMessageTest {
@@ -9,6 +10,6 @@ class ErrorMessageTest {
   @Test
   void classTest() {
     Assertions.assertTrue(
-        Class.forName("com.aftership.sdk.error.ErrorMessage").newInstance() instanceof ErrorMessage);
+        Class.forName("com.aftership.sdk.exception.ErrorMessage").newInstance() instanceof ErrorMessage);
   }
 }
