+++ aftership-sdk/src/test/java/com/aftership/sdk/request/ApiRequestImplJsonBodyTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.Test;
 import java.io.IOException;
 import com.aftership.sdk.AfterShip;
 import com.aftership.sdk.TestUtil;
-import com.aftership.sdk.error.ErrorMessage;
 import com.aftership.sdk.exception.ApiException;
+import com.aftership.sdk.exception.ErrorMessage;
 import com.aftership.sdk.exception.RequestException;
 import com.aftership.sdk.exception.SdkException;
 import okhttp3.mockwebserver.MockWebServer;
